# gravitee-apim-rest-api/gravitee-apim-rest-api-management-v2/gravitee-apim-rest-api-management-v2-rest/src/main/resources/openapi/openapi-apis.yaml
openapi: 3.0.3
info:
    title: Gravitee.io APIM - Management API - APIs
    description: |-
        This is the OpenAPI specification for our new version of APIM Management API.
    contact:
        email: team-apim@graviteesource.com
    license:
        name: Apache 2.0
        url: http://www.apache.org/licenses/LICENSE-2.0.html
    version: 2.0.0
security:
    - BasicAuth: []
    - CookieAuth: []
servers:
    - url: "{protocol}://{managementAPIHost}/management/v2"
      description: APIM Management API v2 - Default base URL
      variables:
          protocol:
              description: The protocol you want to use to communicate with the mAPI
              default: https
              enum:
                  - https
                  - http
          managementAPIHost:
              description: The domain of the server hosting your Management API
              default: localhost:8083
    - url: "{protocol}://{managementAPIHost}/management/v2/organizations/{orgId}"
      description: APIM Management API v2 - Base URL to target specific organizations
      variables:
          protocol:
              description: The protocol you want to use to communicate with the mAPI
              default: https
              enum:
                  - https
                  - http
          managementAPIHost:
              description: The domain of the server hosting your Management API
              default: localhost:8083
          orgId:
              description: The unique ID of your organization
              default: DEFAULT
...
paths:
...
    /environments/{envId}/apis/{apiId}:
        parameters:
            - $ref: "#/components/parameters/envIdParam"
            - $ref: "#/components/parameters/apiIdParam"
        get:
            tags:
                - APIs
            summary: Get a specific API
            description: |-
                Get a specific API.

                User must have at least one management (create, update, or delete) permission on the API to access it.
            operationId: getApi
            responses:
                "200":
                    description: API found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Api"
                default:
                    $ref: "#/components/responses/Error"
        put:
            tags:
                - APIs
            summary: Update an API
            description: |-
                Update a V2, a V4 or a Federated API.

                User must have API_DEFINITION[UPDATE] or API_GATEWAY_DEFINITION[UPDATE] permissions.
            operationId: updateApi
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/UpdateApi"
                required: true
            responses:
                "200":
                    description: API successfully updated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Api"
                default:
                    $ref: "#/components/responses/Error"
        delete:
            parameters:
                - name: closePlans
                  in: query
                  required: false
                  description: |-
                      Set to true to force deletion by closing API plans.<br>
                      ⚠️ Closed plans can't be reopened and every active subscriptions will also be closed. ⚠️
                  schema:
                      type: boolean
            tags:
                - APIs
            summary: Delete an API
            description: |-
                Delete a V2 or a V4 API.

                User must have API_DEFINITION[DELETE] permission.
            operationId: deleteApi
            responses:
                "204":
                    description: API successfully deleted
                default:
                    $ref: "#/components/responses/Error"
...
    /environments/{envId}/apis/{apiId}/plans:
        parameters:
            - $ref: "#/components/parameters/envIdParam"
            - $ref: "#/components/parameters/apiIdParam"
        get:
            parameters:
                - name: statuses
                  in: query
                  required: false
                  description: List of status filters
                  schema:
                      type: array
                      items:
                          $ref: "#/components/schemas/PlanStatus"
                      default: ["published"]
                  explode: false
                - name: securities
                  in: query
                  required: false
                  description: List of security plan type filters
                  schema:
                      type: array
                      items:
                          $ref: "#/components/schemas/PlanSecurityType"
                  explode: false
                - name: mode
                  in: query
                  required: false
                  description: Mode of plan
                  schema:
                      $ref: "#/components/schemas/PlanMode"
                - name: subscribableBy
                  in: query
                  required: false
                  description: Application's identifier. Allow user to get API's plans subscribable by an application.
                  schema:
                      type: string
                - $ref: "#/components/parameters/pageParam"
                - $ref: "#/components/parameters/perPageParam"
            tags:
                - API Plans
            summary: List the API's plans
            description: |
                List plans for a given API, always filtered by a `status` and the possibility to filter by `security`.<br>
                The results are paginated.

                User must have API_PLAN[READ] permissions to access endpoint.
            operationId: listApiPlans
            responses:
                "200":
                    $ref: "#/components/responses/PlansResponse"
                default:
                    $ref: "#/components/responses/Error"
        post:
            tags:
                - API Plans
            summary: Create an API's plan
            description: |-
                Create a new plan on a given API.

                User must have the API_PLAN[CREATE] permission.
            operationId: createApiPlan
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/CreatePlan"
                required: true
            responses:
                "201":
                    description: Plan successfully created
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Plan"
                default:
                    $ref: "#/components/responses/Error"
    /environments/{envId}/apis/{apiId}/plans/{planId}:
        parameters:
            - $ref: "#/components/parameters/envIdParam"
            - $ref: "#/components/parameters/apiIdParam"
            - $ref: "#/components/parameters/planIdParam"
        get:
            tags:
                - API Plans
            summary: Get one API's plan
            description: |-
                Get the API's plan by its identifier.

                User must have the API_PLAN[READ] permission.
            operationId: getApiPlan
            responses:
                "200":
                    description: An API's plan
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Plan"
                default:
                    $ref: "#/components/responses/Error"
        put:
            tags:
                - API Plans
            summary: Update one API's plan
            description: |-
                Update the API's plan.

                User must have the API_PLAN[UPDATE] permission.
            operationId: updateApiPlan
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/UpdatePlan"
                required: true
            responses:
                "200":
                    description: API's plan successfully updated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Plan"
                default:
                    $ref: "#/components/responses/Error"
        delete:
            tags:
                - API Plans
            summary: Delete one API's plan
            description: |-
                Delete the API's plan.

                User must have the API_PLAN[DELETE] permission.
            operationId: deleteApiPlan
            responses:
                "204":
                    description: API's plan successfully deleted
                default:
                    $ref: "#/components/responses/Error"
    /environments/{envId}/apis/{apiId}/plans/{planId}/_close:
        parameters:
            - $ref: "#/components/parameters/envIdParam"
            - $ref: "#/components/parameters/apiIdParam"
            - $ref: "#/components/parameters/planIdParam"
        post:
            tags:
                - API Plans
            summary: Close an API's plan
            description: |-
                Close the API's plan.

                User must have the API_PLAN[UPDATE] permission.
            operationId: closeApiPlan
            responses:
                "200":
                    description: API's plan successfully closed
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Plan"
                default:
                    $ref: "#/components/responses/Error"
    /environments/{envId}/apis/{apiId}/plans/{planId}/_deprecate:
        parameters:
            - $ref: "#/components/parameters/envIdParam"
            - $ref: "#/components/parameters/apiIdParam"
            - $ref: "#/components/parameters/planIdParam"
        post:
            tags:
                - API Plans
            summary: Deprecate an API's plan
            description: |-
                Deprecate the API's plan.

                User must have the API_PLAN[UPDATE] permission.
            operationId: deprecateApiPlan
            responses:
                "200":
                    description: API's plan successfully deprecated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Plan"
                default:
                    $ref: "#/components/responses/Error"
    /environments/{envId}/apis/{apiId}/plans/{planId}/_publish:
        parameters:
            - $ref: "#/components/parameters/envIdParam"
            - $ref: "#/components/parameters/apiIdParam"
            - $ref: "#/components/parameters/planIdParam"
        post:
            tags:
                - API Plans
            summary: Publish an API's plan
            description: |-
                Publish the API's plan.

                User must have the API_PLAN[UPDATE] permission.
            operationId: publishApiPlan
            responses:
                "200":
                    description: API's plan successfully published
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Plan"
                default:
                    $ref: "#/components/responses/Error"

    # Group
...
    /environments/{envId}/categories/{categoryIdOrKey}/apis/{apiId}:
          parameters:
            - $ref: "#/components/parameters/envIdParam"
            - $ref: "#/components/parameters/categoryIdOrKeyParam"
            - $ref: "#/components/parameters/apiIdParam"
          post:
            tags:
              - Categories
            description: Update a specified API within a Category
            operationId: updateCategoryApi
            requestBody:
              content:
                application/json:
                  schema:
                    $ref: "#/components/schemas/UpdateCategoryApi"
              required: true
            responses:
              "200":
                description: The updated CategoryApi
                content:
                  application/json:
                    schema:
                      $ref: "#/components/schemas/CategoryApi"
              default:
                $ref: "#/components/responses/Error"
components:
    schemas:
...
        Api:
            oneOf:
                - $ref: "#/components/schemas/ApiV1"
                - $ref: "#/components/schemas/ApiV2"
                - $ref: "#/components/schemas/ApiV4"
                - $ref: "#/components/schemas/ApiFederated"
            discriminator:
                propertyName: definitionVersion
                mapping:
                    V1: "#/components/schemas/ApiV1"
                    V2: "#/components/schemas/ApiV2"
                    V4: "#/components/schemas/ApiV4"
                    FEDERATED: "#/components/schemas/ApiFederated"
...
        UpdateApi:
            oneOf:
                - $ref: "#/components/schemas/UpdateApiV2"
                - $ref: "#/components/schemas/UpdateApiV4"
                - $ref: "#/components/schemas/UpdateApiFederated"
            discriminator:
                propertyName: definitionVersion
                mapping:
                    V2: "#/components/schemas/UpdateApiV2"
                    V4: "#/components/schemas/UpdateApiV4"
                    FEDERATED: "#/components/schemas/UpdateApiFederated"
...
        PlanSecurityType:
            type: string
            description: Plan security type.
            example: KEY_LESS
            enum:
                - KEY_LESS
                - API_KEY
                - OAUTH2
                - JWT
        PlanStatus:
            type: string
            description: Plan status.
            example: STAGING
            enum:
                - STAGING
                - PUBLISHED
                - DEPRECATED
                - CLOSED
...
        CreatePlan:
            oneOf:
                - $ref: "#/components/schemas/CreatePlanV2"
                - $ref: "#/components/schemas/CreatePlanV4"
            discriminator:
                propertyName: definitionVersion
                mapping:
                    V2: "#/components/schemas/CreatePlanV2"
                    V4: "#/components/schemas/CreatePlanV4"
...
        PlanMode:
            type: string
            description: The behavioural mode of the Plan (Standard for classical plan, Push for subscription plan).
            example: STANDARD
            enum:
                - STANDARD
                - PUSH
...
    parameters:
...
        envIdParam:
            name: envId
            in: path
            required: true
            description: Id or Hrid (Human readable Id) of an environment.
            schema:
                type: string
                default: DEFAULT
        apiIdParam:
            name: apiId
            in: path
            required: true
            description: Id of an API.
            schema:
                type: string
...
        planIdParam:
            name: planId
            in: path
            required: true
            description: Id of a plan.
            schema:
                type: string
...
        pageParam:
            name: page
            in: query
            required: false
            description: The page number for pagination.
            schema:
                type: integer
                default: 1
        perPageParam:
            name: perPage
            in: query
            required: false
            description: |
                The number of items per page for pagination.
            schema:
                type: integer
                default: 10
        ###############################
        # Other Query Parameters      #
        ###############################
...
    responses:
...
        PlansResponse:
            description: Page of API plans
            content:
                application/json:
                    schema:
                        title: "PlansResponse"
                        properties:
                            data:
                                description: List of API plans.
                                type: array
                                items:
                                    $ref: "#/components/schemas/Plan"
                            pagination:
                                $ref: "#/components/schemas/Pagination"
                            links:
                                $ref: "#/components/schemas/Links"
...
        Error:
            description: Generic error response
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/Error"
...
    securitySchemes:
        BasicAuth:
            type: http
            scheme: basic
        CookieAuth:
            type: apiKey
            in: cookie
            name: Auth-Graviteeio-APIM